migrate codebase to postgres database

+++ src/routes/experts/expert.controller.ts
@@ -1,23 +1,42 @@
 import { Request, Response } from "express";
-import Expert from "./expert.model";
+import Expert, { ExpertMap } from "./expert.model";
+import sequelizeDB from "../../config/db";
 import jwt from "jsonwebtoken";
 import bcrypt from "bcrypt";
 import _ from "lodash";
 const fs = require("fs");
 const path = require("path");
 import { appRoot } from "../..";
+import sendEmail from "../../services/email/sendEmail";
+import {
+  verifyEmail,
+  verifyEmailTitle,
+} from "./templates/verifyEmail/verifyEmail";
+import {
+  welcomeEmail,
+  welcomeEmailTitle,
+} from "./templates/welcomeEmail/welcomeEmail";
+import {
+  resetPassword,
+  resetPasswordTitle,
+} from "./templates/resetPassword/resetPassword";
+import {
+  welcomeBack,
+  welcomeBackTitle,
+} from "./templates/welcomeBack/welcomeBack";
 
 class ExpertController {
   async register(req: Request, res: Response) {
     try {
-      const user = await Expert.findOne({ email: req.body.email });
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { email: req.body.email } });
       if (user) {
         return res.status(409).json({
           message: "email already exist",
         });
       }
       const hash = await bcrypt.hash(req.body.password, 10);
-      const newuser = new Expert({
+      const userData = {
         firstname: req.body.firstname,
         lastname: req.body.lastname,
         phone: req.body.phone,
@@ -25,28 +44,42 @@ class ExpertController {
         focusarea: req.body.focusarea,
         havecertifications: req.body.havecertifications,
         password: hash,
-      });
-      newuser
-        .save()
-        .then((response) => {
-          const token = {
-            id: response._id,
-            email: response.email,
-            phone: response.phone,
-          };
+      };
+      await Expert.create(userData)
+        .then((newuser) => {
+          const token: string = jwt.sign(
+            {
+              id: newuser.id,
+              role: newuser.role,
+              email: newuser.email,
+              phone: newuser.phone,
+            },
+            process.env.JWT_SECRET as string
+          );
+          const url = `${process.env.SERVER_URL}/api/${process.env.API_VERSION}/experts/verification/${token}`;
+          sendEmail({
+            to: newuser.email as string,
+            subject: "Prepmeet Email Verification",
+            title: verifyEmailTitle(req.body.firstname),
+            message: verifyEmail(req.body.firstname, url),
+          });
           res.status(201).json({
             message: "success",
             token,
           });
         })
         .catch((err) => {
-          res.status(500).json({
+          res.status(400).json({
             message: "an error occured",
-            error: err,
+            err,
           });
         });
     } catch (error) {
       console.error("user registration error", error);
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
@@ -56,35 +89,46 @@ class ExpertController {
         req.params.token,
         process.env.JWT_SECRET as string
       );
-      const response = await Expert.updateOne(
-        { _id: decodedToken.id },
-        {
-          $set: {
-            emailConfirmed: true,
-          },
-        }
-      );
-      if (response.acknowledged) {
-        // write send mail function here
-        // sendEmail({
-        //   to: decodedToken.email as string,
-        //   subject: "Deonicode: Welcome",
-        //   message: welcomeEmail(decodedToken.username as string),
-        // });
-        res.redirect(301, `${process.env.FRONTEND_URL}`);
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { id: decodedToken.id } });
+      if (user) {
+        user.set({ emailConfirmed: true });
+        await user
+          .save()
+          .then(() => {
+            sendEmail({
+              to: decodedToken.email as string,
+              subject: `Welcome expert ${user.firstname}`,
+              title: welcomeEmailTitle(user.firstname as string),
+              message: welcomeEmail(user.firstname as string),
+            });
+            return res.status(200).json({
+              message: "success",
+            });
+          })
+          .catch((err: any) => {
+            return res.status(500).json({
+              message: "email verification failed",
+              error: err,
+            });
+          });
       } else {
-        res.status(500).json({
-          message: "email verification failed",
+        return res.status(404).json({
+          message: "user not found",
         });
       }
     } catch (error) {
-      console.error("error in email verification", error);
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
   async login(req: Request, res: Response) {
     try {
-      const user = await Expert.findOne({ email: req.body.email });
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { email: req.body.email } });
       if (user) {
         if (user.emailConfirmed === false) {
           return res.status(401).json({
@@ -103,7 +147,8 @@ class ExpertController {
             if (result) {
               const token: string = jwt.sign(
                 {
-                  id: user._id,
+                  id: user.id,
+                  role: user.role,
                   phone: user.phone,
                   email: user.email,
                 },
@@ -126,22 +171,24 @@ class ExpertController {
       }
     } catch (error) {
       console.error("user login error", error);
-      return res.status(500);
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
   async update(req: Request, res: Response) {
-    const user = await Expert.updateOne(
-      {
-        _id: req.params.id,
-      },
-      {
-        $set: {
+    try {
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { id: req.params.id } });
+      if (user) {
+        user.set({
           firstname: req.body.firstname,
           lastname: req.body.lastname,
           phone: req.body.phone,
           email: req.body.email,
-          bip: req.body.bio,
+          bio: req.body.bio,
           education: req.body.education,
           experience: req.body.experience,
           certificates: req.body.certificates,
@@ -166,26 +213,39 @@ class ExpertController {
             IBAN: req.body.IBAN,
             SWIFTBIC: req.body.SWIFTBIC,
           },
-        },
+        });
+        await user
+          .save()
+          .then((resUser) => {
+            const token: string = jwt.sign(
+              {
+                id: resUser?.id,
+                role: resUser?.role,
+                phone: resUser?.phone,
+                email: resUser?.email,
+              },
+              process.env.JWT_SECRET as string
+            );
+            return res.status(200).json({
+              message: "success",
+              token,
+            });
+          })
+          .catch((err: any) => {
+            return res.status(500).json({
+              message: "email verification failed",
+              error: err,
+            });
+          });
+      } else {
+        return res.status(404).json({
+          message: "user not found",
+        });
       }
-    );
-    if (user.acknowledged) {
-      const newuser = await Expert.findOne({ _id: req.params.id });
-      const token: string = jwt.sign(
-        {
-          id: newuser?._id,
-          phone: newuser?.phone,
-          email: newuser?.email,
-        },
-        process.env.JWT_SECRET as string
-      );
-      res.status(200).json({
-        message: "success",
-        token: token,
-      });
-    } else {
-      res.status(404).json({
-        message: "user not found",
+    } catch (error) {
+      return res.status(500).json({
+        message: "an error occured",
+        error,
       });
     }
   }
@@ -209,54 +269,39 @@ class ExpertController {
       });
 
       // Find and delete current image if it exist
-      const user = await Expert.findOne({ _id: req.params.id });
-      if (user?.avatar !== null) {
-        const filePathToDelete = path.join(
-          __dirname,
-          "uploads/expert/profileImages",
-          user?.avatar.doc
-        );
-        // Use fs.unlink to delete the file
-        fs.unlink(appRoot, (err: any) => {
-          if (err) {
-            console.error(`Error deleting file: ${err.message}`);
-          } else {
-            console.log(`File ${filePathToDelete} deleted successfully.`);
-          }
-        });
-      }
-
-      // upload new image to database
-      const profileImage: any[] = Object.entries(files)[0];
-      const image = {
-        doc: profileImage[1].name,
-        key: profileImage[0],
-      };
-
-      const updatedUser = await Expert.updateOne(
-        {
-          _id: req.params.id,
-        },
-        {
-          $set: {
-            avatar: image,
-          },
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { id: req.params.id } });
+      if (user) {
+        if (user?.avatar !== null) {
+          const filePathToDelete = path.join(
+            __dirname,
+            "uploads/expert/profileImages",
+            user.avatar
+          );
+          // Use fs.unlink to delete the file
+          fs.unlink(appRoot, (err: any) => {
+            if (err) {
+              console.error(`Error deleting file: ${err.message}`);
+            } else {
+              console.log(`File ${filePathToDelete} deleted successfully.`);
+            }
+          });
         }
-      );
 
-      if (updatedUser.acknowledged) {
-        const newuser = await Expert.findOne({ _id: req.params.id });
-        const token: string = jwt.sign(
-          {
-            id: newuser?._id,
-            phone: newuser?.phone,
-            email: newuser?.email,
-          },
-          process.env.JWT_SECRET as string
-        );
-        res.status(200).json({
-          message: "success",
-          token: token,
+        // upload new image to database
+        const profileImage: any[] = Object.entries(files)[0];
+        // const image = {
+        //   doc: profileImage[1].name,
+        //   key: profileImage[0],
+        // };
+
+        user.set({
+          avatar: profileImage[1].name,
+        });
+        await user.save().then(() => {
+          return res.status(200).json({
+            message: "success",
+          });
         });
       } else {
         res.status(404).json({
@@ -265,59 +310,72 @@ class ExpertController {
       }
     } catch (error) {
       console.error("error uploading profile image", error);
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
   async updatePassword(req: Request, res: Response) {
-    let user = await Expert.findOne({ _id: req.params.id });
-    if (user) {
-      const { currentPassword, newPassword } = req.body;
-      bcrypt
-        .compare(currentPassword, user.password!)
-        .then((match: any) => {
-          if (match) {
-            bcrypt.hash(newPassword, 10, (error: any, hash: any) => {
-              if (error) {
-                return res.status(500).json({
-                  error: error,
-                });
-              }
-              const passwordUpdate = {
-                password: hash,
-              };
-              user = _.extend(user, passwordUpdate);
-              if (user) {
-                user
-                  .save()
-                  .then((result: any) => {
-                    res.status(200).json({
-                      message: "success",
-                    });
-                  })
-                  .catch((error: any) => {
-                    res.status(500).json({
-                      error: error,
-                    });
+    try {
+      ExpertMap(sequelizeDB);
+      let user = await Expert.findOne({ where: { id: req.params.id } });
+      if (user) {
+        const { currentPassword, newPassword } = req.body;
+        bcrypt
+          .compare(currentPassword, user.password!)
+          .then((match: any) => {
+            if (match) {
+              bcrypt.hash(newPassword, 10, (error: any, hash: any) => {
+                if (error) {
+                  return res.status(500).json({
+                    error: error,
                   });
-              }
+                }
+                const passwordUpdate = {
+                  password: hash,
+                };
+                user = _.extend(user, passwordUpdate);
+                if (user) {
+                  user
+                    .save()
+                    .then((result: any) => {
+                      res.status(200).json({
+                        message: "success",
+                      });
+                    })
+                    .catch((error: any) => {
+                      res.status(500).json({
+                        error: error,
+                      });
+                    });
+                }
+              });
+            } else {
+              return res.status(500).json({
+                message: "passwords do not match",
+              });
+            }
+          })
+          .catch((err: any) => {
+            return res.status(401).json({
+              error: err,
             });
-          } else {
-            return res.status(500).json({
-              message: "passwords do not match",
-            });
-          }
-        })
-        .catch((err: any) => {
-          return res.status(401).json({
-            error: err,
           });
-        });
+      }
+    } catch (error) {
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
   async user(req: Request, res: Response) {
     try {
-      const user = await Expert.findOne({ _id: req.params.id });
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { id: req.params.id } });
       if (user) {
         return res.status(200).json(user);
       } else {
@@ -327,12 +385,17 @@ class ExpertController {
       }
     } catch (error) {
       console.error("error fetching user", error);
+      return res.status(500).json({
+        message: "an error occured",
+        error,
+      });
     }
   }
 
   async users(req: Request, res: Response) {
     try {
-      const users = await Expert.find().sort({ createdAt: -1 });
+      ExpertMap(sequelizeDB);
+      const users = await Expert.findAll();
       if (users) {
         return res.status(200).json(users);
       } else {
@@ -342,16 +405,22 @@ class ExpertController {
       }
     } catch (error) {
       console.error("error fetching users", error);
+      return res.status(500).json({
+        message: "error fetching users",
+        error,
+      });
     }
   }
 
   async forgotPassword(req: Request, res: Response) {
     try {
-      const user = await Expert.findOne({ email: req.body.email });
+      ExpertMap(sequelizeDB);
+      const user = await Expert.findOne({ where: { email: req.body.email } });
       if (user) {
         const resetToken: string = jwt.sign(
           {
-            id: user._id,
+            id: user.id,
+            role: user.role,
             email: user.email,
             phone: user.phone,
           },
@@ -360,13 +429,13 @@ class ExpertController {
             expiresIn: "1h",
           }
         );
-        const url = `${process.env.FRONTEND_URL}/new-password/${resetToken}`;
-        // write send mail function here
-        // sendEmail({
-        //   to: user.email as string,
-        //   subject: "Deonicode: Password Reset",
-        //   message: forgotPasswordEmail(user.username as string, url),
-        // });
+        const url = `${process.env.FRONTEND_URL}/api/${process.env.API_VERSION}/experts/new-password/${resetToken}`;
+        sendEmail({
+          to: user.email as string,
+          subject: "Forgot password - Prepmeet",
+          title: resetPasswordTitle(),
+          message: resetPassword(url),
+        });
         return res.status(200).json({
           message: "success, check your inbox",
         });
@@ -377,12 +446,16 @@ class ExpertController {
       }
     } catch (error) {
       console.error("error in forgot password", error);
+      return res.status(500).json({
+        message: "an error occured",
+      });
     }
   }
 
   async newPassword(req: Request, res: Response) {
     try {
-      let user = await Expert.findOne({ _id: req.params.id });
+      ExpertMap(sequelizeDB);
+      let user = await Expert.findOne({ where: { id: req.params.id } });
       if (user) {
         const { newPassword } = req.body;
         bcrypt.hash(newPassword, 10, async (error: any, hash: any) => {
@@ -402,11 +475,18 @@ class ExpertController {
                 const token: string = jwt.sign(
                   {
                     id: result._id,
+                    role: result.role,
                     phone: result.phone,
                     email: result.email,
                   },
                   process.env.JWT_SECRET as string
                 );
+                sendEmail({
+                  to: result.email as string,
+                  subject: `Welcome back ${result.firstname}`,
+                  title: welcomeBackTitle(result.firstname),
+                  message: welcomeBack(),
+                });
                 res.status(200).json({
                   message: "success",
                   token: token,
@@ -426,6 +506,9 @@ class ExpertController {
       }
     } catch (error) {
       console.error("error in new password", error);
+      return res.status(500).json({
+        message: "an error occured",
+      });
     }
   }
 }
